Add header { authorization: token } on fetchTilList / Add actions/list.js
